Python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain
langchain-openai
langchain-postgres
langchain-community
psycopg[binary,pool]
pypdf
tiktoken
tenacity
# optional UI
streamlit
pymssql
psycopg2-binary
plotly
langgraph
langsmith
fpdf
fpdf2
reportlab
cryptography
pyodbc
pandas
openpyxl
pdf2image
Pillow
azure-ai-documentintelligence==1.0.0b4
openai # Supports Azure OpenAI
python-multipart
loguru
rapidfuzz
phonetics
sqlalchemy
pymysql
PyMuPdf
pymupdf
fuzzywuzzy
jellyfish
aiofiles
python-magic
mcp
mcp[cli]
slowapi
